add RelayRuntimeTypes
